requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
redis[hiredis]==4.6.0
fastapi-cache2[redis]==0.2.1
python-dotenv==1.0.0
pydantic_settings
fasttext-wheel
torch>=1.9.0
sentence-transformers>=2.2.0
# fasttext>=0.9.2
langdetect>=1.0.9
aiohttp>=3.8.0
asyncio>=3.4.3
spacy>=3.5.0
transformers>=4.25.0
nltk>=3.7.0
numpy>=1.21.0